+++ scripts/utils/modify-solidity-coverage-lib-api-js.ts
@@ -7,7 +7,7 @@
 import fs from 'fs';
 import path from 'path';
 
-const apiJsPath = path.join(__dirname, '../node_modules/solidity-coverage/lib/api.js');
+const apiJsPath = path.join(__dirname, '../../node_modules/solidity-coverage/lib/api.js');
 let apiJs = fs.readFileSync(apiJsPath, 'utf8');
 
 apiJs = apiJs.replace('gasPrice = 0x01', 'gasPrice = 0x00');